Moved back stamping integration tests to test_canvas.py as it was originally
